requirements.txt
langchain_core
python-dotenv
langchain_community
langserve
fastapi
uvicorn
sse_starlette
langchainhub
slack_sdk
langchain_groq
langgraph
langchain-huggingface
langchain-pinecone
pinecone
pinecone-text
Ipython